Interfaces for ChaincodeEvent and Status in Java API
Exposing concrete implementation classes that could not be instantiated by client code would make unit testing for client application code harder.
