+++ java/src/main/java/org/hyperledger/fabric/client/Status.java
@@ -6,31 +6,17 @@
 
 package org.hyperledger.fabric.client;
 
-import org.hyperledger.fabric.protos.gateway.CommitStatusResponse;
-
 import static org.hyperledger.fabric.protos.peer.TransactionPackage.TxValidationCode;
 
 /**
  * Information about a transaction that is committed to the ledger.
  */
-public final class Status {
-    private final String transactionId;
-    private final long blockNumber;
-    private final TxValidationCode code;
-
-    Status(final String transactionId, final CommitStatusResponse response) {
-        this.transactionId = transactionId;
-        this.blockNumber = response.getBlockNumber();
-        this.code = response.getResult();
-    }
-
+public interface Status {
     /**
      * Get the transaction ID.
      * @return A transaction ID.
      */
-    public String getTransactionId() {
-        return transactionId;
-    }
+    String getTransactionId();
 
     /**
      * Get the block number in which the transaction committed.
@@ -38,23 +24,17 @@ public final class Status {
      * the number.</p>
      * @return A block number.
      */
-    public long getBlockNumber() {
-        return blockNumber;
-    }
+    long getBlockNumber();
 
     /**
      * Get the committed transaction status code.
      * @return Transaction status code.
      */
-    public TxValidationCode getCode() {
-        return code;
-    };
+    TxValidationCode getCode();
 
     /**
      * Check whether the transaction committed successfully.
      * @return {@code true} if the transaction committed successfully; otherwise {@code false}.
      */
-    public boolean isSuccessful() {
-        return code == TxValidationCode.VALID;
-    };
+    boolean isSuccessful();
 }
